Split local environment integration tests from workspace model
