+++ Tests/QuillCodeAppTests/WorkspaceModelTests.swift
@@ -2458,165 +2458,6 @@ final class WorkspaceModelTests: XCTestCase {
         }))
     }
 
-    func testLocalEnvironmentActionsLoadAndRunFromCommandPaletteIDs() throws {
-        let root = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try "printf local-env-ok".write(
-            to: actionsDirectory.appendingPathComponent("bootstrap-env.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let model = QuillCodeWorkspaceModel()
-        let projectID = model.addProject(path: root, name: "Local Env Project")
-        model.selectProject(projectID)
-
-        let action = try XCTUnwrap(model.selectedProject?.localActions.first)
-        XCTAssertEqual(action.title, "Bootstrap Env")
-        XCTAssertEqual(action.relativePath, ".quillcode/actions/bootstrap-env.sh")
-        XCTAssertTrue(model.runWorkspaceCommand(action.id, workspaceRoot: root))
-
-        let card = try XCTUnwrap(model.currentToolCards.last)
-        XCTAssertEqual(card.title, "host.shell.run")
-        XCTAssertEqual(card.status, .done)
-        let outputJSON = try XCTUnwrap(card.outputJSON)
-        let result = try JSONHelpers.decode(ToolResult.self, from: outputJSON)
-        XCTAssertEqual(result.stdout, "local-env-ok")
-    }
-
-    func testLocalEnvironmentActionMetadataInjectsBoundedEnvironment() throws {
-        let root = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try #"printf "%s|%s|%s|%s" "$QUILL_ENV" "$CACHE_DIR" "$QUOTED_VALUE" "$(printenv BAD-KEY || true)""#.write(
-            to: actionsDirectory.appendingPathComponent("env-check.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Environment Check",
-          "environment": {
-            "QUILL_ENV": "dev",
-            "CACHE_DIR": ".cache/quill",
-            "QUOTED_VALUE": "it's ok",
-            "BAD-KEY": "ignored",
-            "MULTILINE": "bad\\nvalue"
-          }
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("env-check.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let model = QuillCodeWorkspaceModel()
-        let projectID = model.addProject(path: root, name: "Local Env Metadata Project")
-        model.selectProject(projectID)
-
-        let action = try XCTUnwrap(model.selectedProject?.localActions.first)
-        XCTAssertEqual(action.title, "Environment Check")
-        XCTAssertEqual(action.environment, [
-            "CACHE_DIR": ".cache/quill",
-            "QUILL_ENV": "dev",
-            "QUOTED_VALUE": "it's ok"
-        ])
-        XCTAssertEqual(
-            action.command,
-            #"sh '.quillcode/actions/env-check.sh'"#
-        )
-        XCTAssertTrue(model.runWorkspaceCommand(action.id, workspaceRoot: root))
-
-        let card = try XCTUnwrap(model.currentToolCards.last)
-        XCTAssertTrue(card.inputJSON?.contains(#""environment""#) == true)
-        XCTAssertTrue(card.inputJSON?.contains("QUILL_ENV") == true)
-        XCTAssertTrue(card.inputJSON?.contains(ToolCall.redactedEnvironmentValue) == true)
-        XCTAssertFalse(card.inputJSON?.contains(".cache/quill") == true)
-        XCTAssertFalse(card.inputJSON?.contains("it's ok") == true)
-        let outputJSON = try XCTUnwrap(card.outputJSON)
-        let result = try JSONHelpers.decode(ToolResult.self, from: outputJSON)
-        XCTAssertEqual(result.stdout, "dev|.cache/quill|it's ok|")
-    }
-
-    func testLocalEnvironmentActionMetadataRunsFromBoundedWorkingDirectory() throws {
-        let root = try makeTempDirectory()
-        let appDirectory = root.appendingPathComponent("app")
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: appDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try "marker-ok".write(
-            to: appDirectory.appendingPathComponent("marker.txt"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try #"printf "%s|%s" "$(basename "$PWD")" "$(cat marker.txt)""#.write(
-            to: actionsDirectory.appendingPathComponent("cwd-check.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "CWD Check",
-          "workingDirectory": "app"
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("cwd-check.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let model = QuillCodeWorkspaceModel()
-        let projectID = model.addProject(path: root, name: "Local Env CWD Project")
-        model.selectProject(projectID)
-
-        let action = try XCTUnwrap(model.selectedProject?.localActions.first)
-        XCTAssertEqual(action.workingDirectory, "app")
-        XCTAssertEqual(action.command, #"cd 'app' && sh '../.quillcode/actions/cwd-check.sh'"#)
-        XCTAssertTrue(model.runWorkspaceCommand(action.id, workspaceRoot: root))
-
-        let card = try XCTUnwrap(model.currentToolCards.last)
-        let outputJSON = try XCTUnwrap(card.outputJSON)
-        let result = try JSONHelpers.decode(ToolResult.self, from: outputJSON)
-        XCTAssertEqual(result.stdout, "app|marker-ok")
-    }
-
-    func testLocalEnvironmentActionMetadataPassesBoundedTimeout() throws {
-        let root = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try "sleep 2; printf should-not-print".write(
-            to: actionsDirectory.appendingPathComponent("slow.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Slow Check",
-          "timeoutSeconds": 1
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("slow.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let model = QuillCodeWorkspaceModel()
-        let projectID = model.addProject(path: root, name: "Local Env Timeout Project")
-        model.selectProject(projectID)
-
-        let action = try XCTUnwrap(model.selectedProject?.localActions.first)
-        XCTAssertEqual(action.timeoutSeconds, 1)
-        XCTAssertTrue(model.runWorkspaceCommand(action.id, workspaceRoot: root))
-
-        let card = try XCTUnwrap(model.currentToolCards.last)
-        XCTAssertEqual(card.status, .failed)
-        let outputJSON = try XCTUnwrap(card.outputJSON)
-        let result = try JSONHelpers.decode(ToolResult.self, from: outputJSON)
-        XCTAssertFalse(result.ok)
-        XCTAssertEqual(result.error, "Command timed out after 1s.")
-    }
-
     func testEmptyDraftDoesNotCreateThread() async throws {
         let model = QuillCodeWorkspaceModel()
         model.setDraft("   ")
